:truck: chore - Ajustando estrutura do projeto
