+++ Desafio-Total-Shake/src/main/java/br/com/desafio/totalshake/application/controller/response/PedidoDTOResponse.java
@@ -7,7 +7,6 @@ import br.com.desafio.totalshake.domain.model.Status;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 import java.time.LocalDateTime;
 
@@ -52,6 +51,7 @@ public class PedidoDTOResponse {
         return itens;
     }
 
+    @JsonProperty("dataHoraStatus")
     public DataHoraStatusPedido getDataHoraStatusPedido() {
         return dataHoraStatus;
     }
