+++ Desafio-Total-Shake/src/main/java/br/com/desafio/totalshake/domain/service/state/EstadoPedidoFactory.java
@@ -1,8 +1,9 @@
-package br.com.desafio.totalshake.impl;
+package br.com.desafio.totalshake.domain.service.state;
 
 import br.com.desafio.totalshake.domain.model.Pedido;
 import br.com.desafio.totalshake.domain.model.Status;
-import br.com.desafio.totalshake.domain.service.EstadoPedido;
+import br.com.desafio.totalshake.domain.service.state.EstadoPedido;
+import br.com.desafio.totalshake.domain.service.state.impl.*;
 
 public class EstadoPedidoFactory {
