+++ Desafio-Total-Shake/src/main/java/br/com/desafio/totalshake/domain/service/state/impl/CanceladoImpl.java
@@ -1,38 +1,20 @@
-package br.com.desafio.totalshake.impl;
+package br.com.desafio.totalshake.domain.service.state.impl;
 
 import br.com.desafio.totalshake.application.errors.CodInternoErroApi;
 import br.com.desafio.totalshake.application.errors.exceptions.StatusInvalidoException;
 import br.com.desafio.totalshake.domain.model.Pedido;
 import br.com.desafio.totalshake.domain.model.Status;
-import br.com.desafio.totalshake.domain.service.EstadoPedido;
+import br.com.desafio.totalshake.domain.service.state.EstadoPedido;
 
-public class RealizadoImpl implements EstadoPedido {
+public class CanceladoImpl implements EstadoPedido {
 
-    private final Pedido pedido;
-    private final Status realizado = Status.REALIZADO;
+    private Pedido pedido;
+    private final Status cancelado = Status.CANCELADO;
 
-    public RealizadoImpl(Pedido pedido) {
+
+    public CanceladoImpl(Pedido pedido) {
         this.pedido = pedido;
-        this.pedido.setStatus(realizado);
-    }
-
-    @Override
-    public void pagarPedido() {
-        this.pedido.setEstadoPedido(new PagoImpl(this.pedido));
-    }
-
-    @Override
-    public void cancelarPedido() {
-        this.pedido.setEstadoPedido(new CanceladoImpl(this.pedido));
-    }
-
-    @Override
-    public void confirmarPedido() {
-        throw new StatusInvalidoException(
-                CodInternoErroApi.AP301.getMensagem(),
-                CodInternoErroApi.AP301.getCodigo(),
-                realizado
-        );
+        this.pedido.setStatus(cancelado);
     }
 
     @Override
@@ -40,7 +22,34 @@ public class RealizadoImpl implements EstadoPedido {
         throw new StatusInvalidoException(
                 CodInternoErroApi.AP301.getMensagem(),
                 CodInternoErroApi.AP301.getCodigo(),
-                realizado
+                cancelado
+        );
+    }
+
+    @Override
+    public void pagarPedido() {
+        throw new StatusInvalidoException(
+                CodInternoErroApi.AP301.getMensagem(),
+                CodInternoErroApi.AP301.getCodigo(),
+                cancelado
+        );
+    }
+
+    @Override
+    public void confirmarPedido() {
+        throw new StatusInvalidoException(
+                CodInternoErroApi.AP301.getMensagem(),
+                CodInternoErroApi.AP301.getCodigo(),
+                cancelado
+        );
+    }
+
+    @Override
+    public void cancelarPedido() {
+        throw new StatusInvalidoException(
+                CodInternoErroApi.AP301.getMensagem(),
+                CodInternoErroApi.AP301.getCodigo(),
+                cancelado
         );
     }
 
@@ -49,7 +58,7 @@ public class RealizadoImpl implements EstadoPedido {
         throw new StatusInvalidoException(
                 CodInternoErroApi.AP301.getMensagem(),
                 CodInternoErroApi.AP301.getCodigo(),
-                realizado
+                cancelado
         );
     }
 
@@ -58,7 +67,7 @@ public class RealizadoImpl implements EstadoPedido {
         throw new StatusInvalidoException(
                 CodInternoErroApi.AP301.getMensagem(),
                 CodInternoErroApi.AP301.getCodigo(),
-                realizado
+                cancelado
         );
     }
 
@@ -67,7 +76,7 @@ public class RealizadoImpl implements EstadoPedido {
         throw new StatusInvalidoException(
                 CodInternoErroApi.AP301.getMensagem(),
                 CodInternoErroApi.AP301.getCodigo(),
-                realizado
+                cancelado
         );
     }
 
@@ -76,7 +85,7 @@ public class RealizadoImpl implements EstadoPedido {
         throw new StatusInvalidoException(
                 CodInternoErroApi.AP301.getMensagem(),
                 CodInternoErroApi.AP301.getCodigo(),
-                realizado
+                cancelado
         );
     }
 }
