+++ Desafio-Total-Shake/src/test/java/br/com/desafio/totalshake/application/controller/PedidoControllerIntegrationTest.java
@@ -8,9 +8,9 @@ import br.com.desafio.totalshake.domain.model.ItemPedido;
 import br.com.desafio.totalshake.domain.model.Pedido;
 import br.com.desafio.totalshake.domain.model.Status;
 import br.com.desafio.totalshake.domain.repository.PedidoRepository;
-import br.com.desafio.totalshake.impl.CanceladoImpl;
-import br.com.desafio.totalshake.impl.CriadoImpl;
-import br.com.desafio.totalshake.impl.RealizadoImpl;
+import br.com.desafio.totalshake.domain.service.state.impl.CanceladoImpl;
+import br.com.desafio.totalshake.domain.service.state.impl.CriadoImpl;
+import br.com.desafio.totalshake.domain.service.state.impl.RealizadoImpl;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
