+++ Desafio-Total-Shake/src/test/java/br/com/desafio/totalshake/domain/service/state/impl/NaoAutorizadoImplTest.java
@@ -1,8 +1,9 @@
-package br.com.desafio.totalshake.impl;
+package br.com.desafio.totalshake.domain.service.state.impl;
 
 import br.com.desafio.totalshake.application.errors.exceptions.StatusInvalidoException;
 import br.com.desafio.totalshake.domain.model.Pedido;
 import br.com.desafio.totalshake.domain.model.Status;
+import br.com.desafio.totalshake.domain.service.state.impl.NaoAutorizadoImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
