+++ Desafio-Total-Shake/src/test/java/br/com/desafio/totalshake/domain/service/state/impl/ProntoImplTest.java
@@ -1,8 +1,10 @@
-package br.com.desafio.totalshake.impl;
+package br.com.desafio.totalshake.domain.service.state.impl;
 
 import br.com.desafio.totalshake.application.errors.exceptions.StatusInvalidoException;
 import br.com.desafio.totalshake.domain.model.Pedido;
 import br.com.desafio.totalshake.domain.model.Status;
+import br.com.desafio.totalshake.domain.service.state.impl.ProntoImpl;
+import br.com.desafio.totalshake.domain.service.state.impl.SaiuParaEntregaImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
